Factor out parsing of component_norm to StrategiesParsing
