+++ src/solver/src/strategies_parsing.C
@@ -27,6 +27,7 @@
 
 // libMesh
 #include "libmesh/getpot.h"
+#include "libmesh/system_norm.h"
 
 namespace GRINS
 {
@@ -50,4 +51,20 @@ namespace GRINS
     return input("unsteady-solver/max_growth", 0.0 );
   }
 
+  void StrategiesParsing::parse_component_norm( const GetPot& input, libMesh::SystemNorm& component_norm )
+  {
+    const unsigned int n_component_norm =
+      input.vector_variable_size("unsteady-solver/component_norm");
+    for (unsigned int i=0; i != n_component_norm; ++i)
+      {
+        const std::string current_norm = input("component_norm", std::string("L2"), i);
+        // TODO: replace this with string_to_enum with newer libMesh
+        if (current_norm == "GRINSEnums::L2")
+          component_norm.set_type(i, libMesh::L2);
+        else if (current_norm == "GRINSEnums::H1")
+          component_norm.set_type(i, libMesh::H1);
+        else
+          libmesh_not_implemented();
+      }
+  }
 } // end namespace GRINS
